testing-v2/scenarios/gaming-leaderboard/iterations/iteration-002-python/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
azure-cosmos>=4.7.0
aiohttp>=3.9.0
pydantic>=2.5.0
